+++ storage/execout/config.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/streamingfast/bstream"
 	"github.com/streamingfast/derr"
 	"github.com/streamingfast/dstore"
 	"go.uber.org/zap"
@@ -60,12 +59,15 @@ func (c *Config) Name() string                        { return c.name }
 func (c *Config) ModuleKind() pbsubstreams.ModuleKind { return c.modKind }
 func (c *Config) ModuleInitialBlock() uint64          { return c.moduleInitialBlock }
 
-func (c *Config) ListSnapshotFiles(ctx context.Context, inRange *bstream.Range) (files FileInfos, err error) {
+func (c *Config) ListSnapshotFiles(ctx context.Context, from uint64, to uint64) (files FileInfos, err error) {
 	err = derr.RetryContext(ctx, 3, func(ctx context.Context) error {
 		// We must reset accumulated files between each retry
 		files = nil
 
-		return c.objStore.WalkFrom(ctx, "", computeDBinFilename(inRange.StartBlock(), 0), func(filename string) (err error) {
+		fromFilename := computeDBinFilename(from, 0)
+		toFilename := computeDBinFilename(to, 0)
+
+		err := c.objStore.WalkFromTo(ctx, "", fromFilename, toFilename, func(filename string) (err error) {
 			var fileInfo *FileInfo
 
 			switch c.modKind {
@@ -80,13 +82,13 @@ func (c *Config) ListSnapshotFiles(ctx context.Context, inRange *bstream.Range) 
 				c.logger.Warn("seen exec output file that we don't know how to parse", zap.String("filename", filename), zap.Error(err))
 				return nil
 			}
-			if inRange.ReachedEndBlock(fileInfo.BlockRange.ExclusiveEndBlock - 1) {
-				return dstore.StopIteration
-			}
-
 			files = append(files, fileInfo)
 			return nil
 		})
+		if err == dstore.StopIteration {
+			return nil
+		}
+		return err
 	})
 	if err != nil {
 		return nil, fmt.Errorf("walking files: %s", err)