+++ docs/release-notes/change-log.md
@@ -17,6 +17,8 @@ Introduces Global Worker Pool management, updating dependencies, and improving t
 * Added 'shared cache' on tier1: execution of modules near the HEAD of the chain will be done once for a given module hash and the result shared between requests.
   This will reduce CPU usage and increase performance when many requests are using the same modules (ex: foundational modules)
 
+* Improved "time to first block" when a lot of cached files exist on dependency substreams modules by skipping reads segments that won't be used and assuming stores "full KVs" are always filled sequentially (since they are!)
+
 * Limit parallel execution of a stage's layer.
 
   Previously, the engine was executing modules in a stage's layer all in parallel. We now change that behavior, development mode will from now on execute every sequentially and when in production mode will limit parallelism to 2 (hard-coded) for now.